Move test beans to common module
