+++ sdk-quarkus/src/test/resources/application.properties
@@ -1,1 +1,1 @@
-quarkus.arc.selected-alternatives=io.a2a.server.apps.quarkus.TestHttpClient
+quarkus.arc.selected-alternatives=io.a2a.server.apps.common.TestHttpClient
